pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elasticsearch<8.0.0",
    "faiss-cpu>=1.13.2",
    "gradio>=6.13.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.14",
    "langchain-openrouter>=0.2.1",
    "minisearch>=1.0.1",
    "openai>=2.32.0",
    "pypdf>=6.10.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "tdqm>=0.0.1",
    "tiktoken>=0.12.0",
]
